pyproject.toml
[tool.poetry]
packages = [{include = "kiwoom"}]

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-mock = "^3.12.0"
black = "^24.4.2"
flake8 = "^7.0.0"
isort = "^5.13.2"
mypy = "^1.10.0"

[project]
name = "python-kiwoom"
version = "0.1.0"
description = "A Python wrapper for the Kiwoom REST API."
authors = [{name = "Your Name", email = "you@example.com"}]
license = "MIT"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "httpx==0.27.0",
    "pydantic==2.7.1",
    "websockets==12.0",
    "python-dotenv==1.0.0",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
